third commit

+++ psg.cpp
@@ -117,7 +117,7 @@ void PSG::MakeEnvelopTable()
 		2,2, 2,0, 2,1, 2,3, 1,1, 1,3, 1,2, 1,0,
 	};
 	static uint8_t table2[4] = {  0,  0, 31, 31 };
-	static uint8_t table3[4] = {  0,  1, -1,  0 };
+	static uint8_t table3[4] = {  0,  1, (uint8_t)-1,  0 };
 
 	uint32_t* ptr = enveloptable[0];
